+++ src/main/kotlin/di/cliModule.kt
@@ -1,4 +1,4 @@
-package di.modules
+package di
 
 import org.koin.dsl.module
 import presentation.PlanMateCLI
